requirements.txt
Django==1.9.7
Pillow==3.1.1
django-resized==0.3.5
factory_boy==2.7.0